Added meta stream
Added function cairo_ps_surface_add_meta_stream which can be used to
write a file which will contain byte offsets of the different sections of
the created PostScript file.

Added a unique prefix to all PostScript resource names, which makes it
easier to handle / merge PostScript output created by cairo.
The unique prefix consists of the pid and the current timestamp.